Consolidate duplicate Series panels in post editor (#3)
Replace the custom PluginDocumentSettingPanel with an extension of the
default WordPress taxonomy panel using the editor.PostTaxonomyType filter.

This eliminates the duplicate panel issue where both WordPress's auto-generated
taxonomy panel and our custom panel appeared in the sidebar.

- Add series-taxonomy-filter.tsx with HOC to extend the term selector
- Add series-extension-fields.tsx for order and short title fields
- Update index.ts to use addFilter instead of registerPlugin
- Remove series-panel.tsx (no longer needed)
