+++ src/sidebar/index.ts
@@ -1,15 +1,18 @@
 /**
- * Content Series - Editor Sidebar Panel
+ * Content Series - Editor Sidebar Extension
  *
- * Registers a plugin that adds a Series panel to the post editor sidebar.
+ * Extends the default Series taxonomy panel with additional fields
+ * (order in series, short title) using the editor.PostTaxonomyType filter.
  */
 
-import { registerPlugin } from '@wordpress/plugins';
-import SeriesPanel from './series-panel';
+import { addFilter } from '@wordpress/hooks';
+
+import { extendSeriesTermSelector } from './series-taxonomy-filter';
 
 import './style.scss';
 
-registerPlugin( 'content-series-sidebar', {
-	render: SeriesPanel,
-	icon: 'list-view',
-} );
+addFilter(
+	'editor.PostTaxonomyType',
+	'content-series/extend-series-selector',
+	extendSeriesTermSelector
+);
